NLP padding
Padding as we know all the neural networks need to have inputs that should be in similar shape and size. When we pre-process the texts and use the texts as an input for our Model. Note that not all the sequences have the same length, as we can say naturally some of the sequences are long in length and some are short. Where we know that we need to have the inputs of the same size, nowhere padding comes into the picture. The inputs should be in the same size at that time padding is necessary.